feature/VIH-6438 : Fix for muting
