+++ VideoWeb/VideoWeb.Contract/Responses/ParticipantForUserResponse.cs
@@ -1,58 +1,7 @@
-using System;
-using VideoWeb.Common.Models;
-
 namespace VideoWeb.Contract.Responses
 {
-    public class ParticipantForUserResponse
+    public class ParticipantForUserResponse : ParticipantResponse
     {
-        /// <summary>
-        /// The participant Id
-        /// </summary>
-        public Guid Id { get; set; }
         
-        /// <summary>
-        /// The participant username
-        /// </summary>
-        public string Username { get; set; }
-        
-        /// <summary>
-        /// The participant display name during a conference
-        /// </summary>
-        public string DisplayName { get; set; }
-
-        /// <summary>
-        /// The participant first name
-        /// </summary>
-        public string FirstName { get; set; }
-
-        /// <summary>
-        /// The participant last name
-        /// </summary>
-        public string LastName { get; set; }
-        
-        /// <summary>
-        /// The participant role in conference
-        /// </summary>
-        public Role Role { get; set; }
-        
-        /// <summary>
-        /// The current status of a participant
-        /// </summary>
-        public ParticipantStatus Status { get; set; }
-        
-        /// <summary>
-        /// The representee (if participant is a representative)
-        /// </summary>
-        public string Representee { get; set; }
-        
-        /// <summary>
-        /// The group a participant belongs to
-        /// </summary>
-        public string CaseTypeGroup { get; set; }
-        
-        /// <summary>
-        /// The display name provided to the pexip node (includes tile position, display name and id)
-        /// </summary>
-        public string PexipDisplayName {get; set; }
     }
 }
