+++ VideoWeb/VideoWeb/Mappings/ConferenceResponseMapper.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using VideoWeb.Common.Models;
 using VideoWeb.Contract.Responses;
+using VideoWeb.Helpers;
 using VideoWeb.Services.Video;
 using UserRole = VideoWeb.Services.Video.UserRole;
 
@@ -46,47 +47,9 @@ namespace VideoWeb.Mappings
             response.PexipNodeUri = conference.Meeting_room.Pexip_node;
             response.PexipSelfTestNodeUri = conference.Meeting_room.Pexip_self_test_node;
 
-            var tiledParticipants = conference.Participants.Where(x =>
-             x.User_role == UserRole.Individual || x.User_role == UserRole.Representative).ToList();
-            if (tiledParticipants.Count > 4)
-            {
-                // If the number of participants is more than 4, then simply increment the tile numbers
-                var position = 1;
-                foreach (var participant in response.Participants)
-                {
-                    if (participant.Role == Role.Judge)
-                    {
-                        participant.TiledDisplayName = GetTiledDisplayName(participant, 0);
-                    }
-                    else
-                    {
-                        participant.TiledDisplayName = GetTiledDisplayName(participant, position);
-                        position++;
-                    }
-                }
-            }
-            else
-            {
-                var partyGroups = tiledParticipants.GroupBy(x => x.Case_type_group).ToList();
-                foreach (var group in partyGroups)
-                {
-                    var pats = group.ToList();
-                    var position = partyGroups.IndexOf(group) + 1;
-                    foreach (var p in pats)
-                    {
-                        var participant = response.Participants.Find(x => x.Id == p.Id);
-                        participant.TiledDisplayName = GetTiledDisplayName(participant, position);
-                        position += 2;
-                    }
-                }
-            }
+            ParticipantTilePositionHelper.AssignTilePositions(response.Participants);
 
             return response;
-        }
-
-        private static string GetTiledDisplayName(ParticipantResponse participant, int position)
-        {
-            return  $"T{position};{participant.DisplayName};{participant.Id}";
         }
     }
 }
